% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_savedmodel}
\alias{load_savedmodel}
\title{Load a SavedModel}
\usage{
load_savedmodel(sess, model_dir = NULL)
}
\arguments{
\item{sess}{The TensorFlow session.}

\item{model_dir}{The path to the exported model, as a string.}
}
\description{
Loads a SavedModel using the given TensorFlow session and
returns the model's graph.
}
