% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnist.R
\name{mnist_train_save}
\alias{mnist_train_save}
\title{Trains and Save MNIST}
\usage{
mnist_train_save(model_path)
}
\arguments{
\item{model_path}{Destination path where the model will be saved.}
}
\description{
Trains and saves MNIST using TensorFlow
}
