% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cloudml.R
\name{predict_savedmodel.cloudml_prediction}
\alias{predict_savedmodel.cloudml_prediction}
\title{Predict using a CloudML SavedModel}
\usage{
\method{predict_savedmodel}{cloudml_prediction}(instances, model = NULL,
  signature_name = "serving_default", version = NULL, ...)
}
\arguments{
\item{instances}{A list of prediction instances to be passed as input tensors
to the service. Even for single predictions, a list with one entry is expected.}

\item{model}{The model as a local path, REST url, CloudML name or graph object.}

\item{signature_name}{The named entry point to use in the model for prediction.}

\item{version}{The version of the CloudML model.}

\item{...}{Additional arguments. See \code{?predict_savedmodel.cloudml_prediction}
and \code{?predict_savedmodel.graph_prediction} for additional options.}
}
\description{
Performs a prediction using a CloudML model.
}
