% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_savedmodel}
\alias{predict_savedmodel}
\title{Predicts over a Saved Model}
\usage{
predict_savedmodel(input, location, service = c("cloudml", "file", "api"),
  signature_name = tf$saved_model$signature_constants$DEFAULT_SERVING_SIGNATURE_DEF_KEY)
}
\arguments{
\item{input}{The prediction input to be passed as input tensors to the service.}

\item{location}{The location as a local path, REST url or CloudML name.}

\item{service}{The type of service, valid values are \code{cloudml},
\code{file}, \code{api}.}

\item{signature_name}{The named entry point to use in the model for prediction.}
}
\description{
Runs a prediction over a saved model file, local service or cloudml model.
}
