% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_webapi.R
\name{predict_savedmodel.webapi_prediction}
\alias{predict_savedmodel.webapi_prediction}
\title{Predict using a Web API}
\usage{
\method{predict_savedmodel}{webapi_prediction}(instances,
  model = "http://127.0.0.1:8089/api/serving_default/predict/",
  signature_name = "serving_default", ...)
}
\description{
Performs a prediction using a Web API providing a SavedModel.
}
