% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{export_savedmodel}
\alias{view_savedmodel}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{tensorflow}{\code{\link[tensorflow]{export_savedmodel}}, \code{\link[tensorflow]{view_savedmodel}}}
}}

