% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_export.R
\name{predict_savedmodel.export_prediction}
\alias{predict_savedmodel.export_prediction}
\title{Predict using an Exported SavedModel}
\usage{
\method{predict_savedmodel}{export_prediction}(instances, model = NULL,
  signature_name = "serving_default", ...)
}
\description{
Performs a prediction using a locally exported SavedModel.
}
