% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_graph.R
\name{predict_savedmodel.graph_prediction}
\alias{predict_savedmodel.graph_prediction}
\title{Predict using a Loaded SavedModel}
\usage{
\method{predict_savedmodel}{graph_prediction}(instances,
  signature_name = "serving_default", sess, ...)
}
\arguments{
\item{sess}{The active TensorFlow session.}
}
\description{
Performs a prediction using a SavedModel model already loaded using
\code{load_savedmodel()}.
}
