% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-interface.R
\name{predict_savedmodel}
\alias{predict_savedmodel}
\title{Predicts over a Saved Model}
\usage{
predict_savedmodel(instances, location, service = c("export", "cloudml",
  "webapi"),
  signature_name = tf$saved_model$signature_constants$DEFAULT_SERVING_SIGNATURE_DEF_KEY,
  ...)
}
\arguments{
\item{instances}{A list of prediction instances to be passed as input tensors
to the service. Even for single predictions, a list with one entry is expected.}

\item{location}{The location as a local path, REST url or CloudML name.}

\item{service}{The type of service, valid values are \code{cloudml},
\code{export}, \code{webapi}.}

\item{signature_name}{The named entry point to use in the model for prediction.}

\item{...}{Additional arguments, currently not in use.}
}
\description{
Runs a prediction over a saved model file, local service or cloudml model.
}
