% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-interface.R
\name{predict_savedmodel}
\alias{predict_savedmodel}
\title{Predict using a SavedModel}
\usage{
predict_savedmodel(instances, model = NULL, type = c("export", "cloudml",
  "webapi", "graph"), signature_name = "serving_default", ...)
}
\arguments{
\item{instances}{A list of prediction instances to be passed as input tensors
to the service. Even for single predictions, a list with one entry is expected.}

\item{model}{The model as a local path, REST url, CloudML name or graph object.}

\item{type}{The type of object performing the prediction. Valid values
are \code{cloudml}, \code{export}, \code{webapi} and \code{graph}.}

\item{signature_name}{The named entry point to use in the model for prediction.}

\item{...}{Additional arguments, currently not in use.}
}
\description{
Runs a prediction over a saved model file, local service or cloudml model.
}
